Add the mock endpoint to the library
It's extremely hard to generalise this the way you'd like. You want an
object that generates things that can be iterated over: it's not valid
in the iterator contract for the lifetime of the items to be bound to
the lifetime of the iterator, so there has to be another object.